Fix conflicting data types

+++ bothub_nlp_api/handlers/intent_sentence_suggestion.py
@@ -34,7 +34,7 @@ def _intent_sentence_suggestion(
         raise ValidationError("Invalid number of sentences to generate")
     if (
         not percentage_to_replace
-        or type(percentage_to_replace) != int
+        or type(percentage_to_replace) != float
         or percentage_to_replace <= 0
         or percentage_to_replace > 100
     ):
